Updated to use Zulu time and ISO 8601 format.
The timestamps were generated with second accuracy from `time` and
converted to strings using `asctime`. Unfortunately the accuracy wasn't
enough and the string format didn't take time zones into account, so
this was changed. Now the times are generated with `gettimeofday` for
at least millisecond accuracy, although the function claims microsecond
accuracy if the CPU and RTC can deliver it. The time is shifted to Zulu
with `gmtime` and converted to a string with `strftime` so the ISO 8601
format can be used.